fix: update news widget
