+++ .env.sample
@@ -1,1 +1,2 @@
 VENLY_ID=Testaccount
+VITE_NEWS_KEY=
