+++ src/constants/index.ts
@@ -34,7 +34,7 @@ export const SQUID_API = 'https://api.0xsquid.com';
 export const PANGOLIN_TOKENS_REPO_RAW_BASE_URL = `https://raw.githubusercontent.com/pangolindex/tokens`;
 export const ANALYTICS_PAGE = 'https://info.pangolin.exchange';
 export const PANGOLIN_API_BASE_URL = `https://api.pangolin.exchange`;
-export const DIRECTUS_URL_NEWS = `https://pangolin.directus.app`;
+export const NEWS_API_URL = `https://pangolin.hasura.app/v1/graphql`;
 export const OPEN_API_DEBANK = 'https://api.debank.com/';
 export const LANDING_PAGE = 'https://pangolin.exchange';
 export const NEAR_API_BASE_URL = `https://testnet-indexer.ref-finance.com`;
